Fix realtime time differences.
Discrepancy caused by the app showing departure times, but the
realtime data is based on arrival times, and sometimes they differ.
This was introducing a small error into the calculated realtime
differences. Stop 1907 is currently a good test.
Also was not correctly setting the displayed field to blank when
there is no RT data, so it was sometimes displaying an old string.